// packages/twenty-server/src/engine/core-modules/auth/types/jwt-token-type.enum.ts
export enum JwtTokenTypeEnum {
  ACCESS = 'ACCESS',
  REFRESH = 'REFRESH',
  WORKSPACE_AGNOSTIC = 'WORKSPACE_AGNOSTIC',
  LOGIN = 'LOGIN',
  FILE = 'FILE',
  FILE_UPLOAD = 'FILE_UPLOAD',
  API_KEY = 'API_KEY',
  REMOTE_SERVER = 'REMOTE_SERVER',
  KEY_ENCRYPTION_KEY = 'KEY_ENCRYPTION_KEY',
  APPLICATION_ACCESS = 'APPLICATION_ACCESS',
  APPLICATION_REFRESH = 'APPLICATION_REFRESH',
  APP_OAUTH_STATE = 'APP_OAUTH_STATE',
  MCP_CONNECTION_OAUTH_STATE = 'MCP_CONNECTION_OAUTH_STATE',
  APPLICATION_REGISTRATION_GITHUB_CLAIM_STATE = 'APPLICATION_REGISTRATION_GITHUB_CLAIM_STATE',
  APPROVED_ACCESS_DOMAIN = 'APPROVED_ACCESS_DOMAIN',
  PLAYGROUND = 'PLAYGROUND',
}